Fix issue resulting in 1 byte files for "Save Plugin as Folder"
What was happening.  When attempting to save a plugin as a folder, non-updated shadow tiddlers would render as 1 byte files.  

Why: The shadow tiddler was not present in the wiki object on the server side.
Why: The plugin tiddler was never unpacked.
Why: In ServerSide.js non $:/core tiddlers are added to the wiki object after the call to `.registerPluginTiddlers`.

Fix: Move `.loadWikiTiddlers` before the call to `.registerPluginTiddlers`.