fix save throttling and localization handling
Languages are now handled correctly, they are added to the wiki the same way as plugins so they can be added to a wiki and then set like normal

Bob no longer tries to throttle saving, it uses the new(ish) updates to the internal syncer to make it unnecessary.
This makes saving a lot of tiddlers at once much faster.